+++ components/ui/Footer.tsx
@@ -1,0 +1,66 @@
+import React from 'react';
+import { Phone, Mail, MapPin } from 'lucide-react';
+
+const Footer = () => {
+  return (
+    <footer className="bg-gray-900 text-white py-12">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+        <div className="grid md:grid-cols-4 gap-8">
+          <div>
+            <div className="flex items-center space-x-3 mb-4">
+              <div className="w-10 h-10 bg-blue-600 rounded-lg flex items-center justify-center">
+                <span className="text-white font-bold">MCA</span>
+              </div>
+              <span className="font-bold">Ministry of Corporate Affairs</span>
+            </div>
+            <p className="text-gray-400">
+              Government of India initiative for empowering youth through practical governance experience.
+            </p>
+          </div>
+          
+          <div>
+            <h3 className="font-semibold mb-4">Quick Links</h3>
+            <ul className="space-y-2 text-gray-400">
+              <li><a href="#" className="hover:text-white">About Programme</a></li>
+              <li><a href="#" className="hover:text-white">Eligibility</a></li>
+              <li><a href="#" className="hover:text-white">Application Process</a></li>
+              <li><a href="#" className="hover:text-white">FAQs</a></li>
+            </ul>
+          </div>
+          
+          <div>
+            <h3 className="font-semibold mb-4">Resources</h3>
+            <ul className="space-y-2 text-gray-400">
+              <li><a href="#" className="hover:text-white">Guidelines</a></li>
+              <li><a href="#" className="hover:text-white">Forms & Documents</a></li>
+              <li><a href="#" className="hover:text-white">Previous Year Data</a></li>
+              <li><a href="#" className="hover:text-white">Success Stories</a></li>
+            </ul>
+          </div>
+          
+          <div>
+            <h3 className="font-semibold mb-4">Contact Info</h3>
+            <div className="space-y-2 text-gray-400">
+              <div className="flex items-center">
+                <Phone className="w-4 h-4 mr-2" />
+                <span>+91-11-2338-4000</span>
+              </div>
+              <div className="flex items-center">
+                <Mail className="w-4 h-4 mr-2" />
+                <span>internship@mca.gov.in</span>
+              </div>
+              <div className="flex items-start">
+                <MapPin className="w-4 h-4 mr-2 mt-1" />
+                <span>A-Wing, Shastri Bhawan, New Delhi - 110001</span>
+              </div>
+            </div>
+          </div>
+        </div>
+        
+        <div className="border-t border-gray-800 mt-8 pt-8 text-center text-gray-400">
+          <p>&copy; 2025 Ministry of Corporate Affairs, Government of India. All rights reserved.</p>
+        </div>
+      </div>
+    </footer>
+  );
+};
